+++ main.go
@@ -3,7 +3,6 @@ package main
 import (
 	"log"
 	"main/api/countryData"
-	geocoords "main/api/geoCoords"
 	"main/api/holidays"
 	"main/db"
 	"main/dict"
@@ -30,8 +29,6 @@ func main() {
 	}
 	//set URL with port
 	dict.URL = dict.URL + ":" + port
-
-	http.HandleFunc("/weather-rest/v1/geocoord/", geocoords.MethodHandler)
 
 	//Get all countries endpoint:
 	http.HandleFunc("/weather-rest/v1/restCountries/", countryData.HandleRestCountry)